backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
yfinance==0.2.65
backtrader==1.9.78.123
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.9.2